+++ CHANGELOG.md
@@ -19,6 +19,9 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 
 - Async method `Container.aget()`.
   This was necessary for generator-based cleanups.
+  It works with sync factories too, so you can use it universally in async code.
+- Async method `ServicePing.aping()`.
+  It works with sync factories and pings too, so you can use it universally in async code.
 
 
 ### Changed